new file:   namedtuplefetchall.py 	new file:   templates/NearBeach/task_information/task_assigned_users.html 	new file:   templates/NearBeach/task_information/task_costs.html 	new file:   templates/NearBeach/task_information/task_customers.html 	new file:   templates/NearBeach/task_information/task_document_folders.html 	new file:   templates/NearBeach/task_information/task_documents.html 	new file:   templates/NearBeach/task_information/task_history.html 	modified:   views.py 	modified:   views_project_information.py 	new file:   views_task_information.py
