+++ namedtuplefetchall.py
@@ -1,0 +1,11 @@
+# Extra functionality
+"""
+The following function helps change the cursor's results into useable
+SQL that the html templates can read.
+"""
+from collections import namedtuple
+def namedtuplefetchall(cursor):
+    "Return all rows from a cursor as a namedtuple"
+    desc = cursor.description
+    nt_result = namedtuple('Result', [col[0] for col in desc])
+    return [nt_result(*row) for row in cursor.fetchall()]
